// src/app/packages/interfaces/Package.ts
export interface IPackage{
    id_paquete:         string
    peso:               string
    estado:             string
    direccion:          string
    poblacion:          string
    nombre_de_afiliado: string
    codigo_de_afiliado: string
    telefono:           string
    celular:            string
    detalles:           string
}